feat(key-result): finishes owner select
