+++ src/components/KeyResult/Single/Sections/Owner/queries.gql
@@ -1,23 +1,3 @@
-query GET_USER_LIST {
-  users(first: 1000) {
-    edges {
-      node {
-        id
-        fullName
-        picture
-        teams {
-          edges {
-            node {
-              id
-              name
-            }
-          }
-        }
-      }
-    }
-  }
-}
-
 mutation UPDATE_KEY_RESULT_OWNER($keyResultID: ID!, $userID: ID!) {
   updateKeyResult(id: $keyResultID, data: {
     ownerId: $userID
